backend/requirements.txt
fastapi[all]>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.22
pypdf2>=3.0.0
tiktoken>=0.6.0
langgraph>=0.0.20
onnxruntime>=1.19.0
python-dotenv>=1.0.0